This function takes a positive integer n as input and returns the sum of all that number's digits.
The purpose of this function is to sum the input

If ~input~ == int; the following will occur:

input is converted into a string then converted into a list
then is stored in a variable = num

the then stored variable is converted back into an int using a for
iteration.

the result is then returned into a sum() function which will sum the
individual integers.